Use relative markdown link to collection README.md
While the /README.md notation works for GitHub it may not work for other
git platforms / web-interfaces.

As it is considered best practice to locally install the collection
into the folder of the root playbook that link should be as compatible
as possible.

Additionally a period is appended to the end of the sentence.